refactor. ArenaProcessingService -> ArenaEventProcessingService
Litt klarere navnebruk for service som prosesserer arena events.
Det gjør det litt enklere når ny service kommer for å opprette avtaler.
